Added ResponseHandler class, and some refactoring

+++ src/main/java/es/users/controllers/UserController.java
@@ -3,16 +3,16 @@ package es.users.controllers;
 import java.util.Optional;
 
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
 
-import es.users.Constants;
 import es.users.api.UserApiDelegate;
 import es.users.dto.UserRequest;
-import es.users.dto.UserResponse;
+import es.users.handler.ResponseHandler;
+import es.users.records.UserResponse;
 import es.users.services.implementations.UserService;
+import es.users.util.Utils;
 import io.swagger.v3.oas.annotations.parameters.RequestBody;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
@@ -31,58 +31,50 @@ public class UserController implements UserApiDelegate {
 
 
     @Override
-    public ResponseEntity<UserResponse> saveUser(@Valid @RequestBody UserRequest userRequest) {
-        String acceptStr = request.getHeader(Constants.ACCEPT_STRING);
-        return switch (acceptStr) {
-            case MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_XML_VALUE -> {
-                UserResponse userResponse = userService.createUser(userRequest);
-                yield new ResponseEntity<>(userResponse, HttpStatus.OK);
-            }
-            default -> new ResponseEntity<>(new UserResponse(), HttpStatus.NOT_ACCEPTABLE);
-        };
+    public ResponseEntity<Object> saveUser(@Valid @RequestBody UserRequest userRequest) {
+        if (Utils.isInvalidAcceptHeader(request)) {
+            return ResponseHandler.notAcceptableResponse();
+        }
+        UserResponse userResponse = userService.createUser(userRequest);
+        return ResponseHandler.buildResponse("User created successfully!", HttpStatus.OK, userResponse);
     }
 
 
     @Override
-    public ResponseEntity<UserResponse> getUserById(@PathVariable("userId") Long userId) {
-        String acceptStr = request.getHeader(Constants.ACCEPT_STRING);
-        return switch (acceptStr) {
-            case MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_XML_VALUE -> {
-                Optional<UserResponse> userResponse = userService.getUserById(userId);
-                yield userResponse.map(usr -> new ResponseEntity<>(usr, HttpStatus.OK))
-                        .orElseGet(() -> new ResponseEntity<>(new UserResponse(), HttpStatus.NOT_FOUND));
-            }
-            default -> new ResponseEntity<>(new UserResponse(), HttpStatus.NOT_ACCEPTABLE);
-        };
+    public ResponseEntity<Object> getUserByNif(@PathVariable("userNif") String userNif) {
+        if (Utils.isInvalidAcceptHeader(request)) {
+            return ResponseHandler.notAcceptableResponse();
+        }
+        Optional<UserResponse> userResponse = userService.getUserByNif(userNif);
+        return userResponse
+                .map(usr -> ResponseHandler.buildResponse("User retrieved successfully.", HttpStatus.OK, usr))
+                .orElseGet(() -> ResponseHandler.buildResponse("User not found!", HttpStatus.NOT_FOUND,
+                        new UserResponse("", "", "")));
     }
 
 
     @Override
-    public ResponseEntity<UserResponse> updateUserByNif(@PathVariable("nif") String nif,
+    public ResponseEntity<Object> updateUserByNif(@PathVariable("nif") String nif,
             @Valid @RequestBody UserRequest userRequest) {
-        String acceptStr = request.getHeader(Constants.ACCEPT_STRING);
-        return switch (acceptStr) {
-            case MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_XML_VALUE -> {
-                Optional<UserResponse> userResponse = userService.updateUserNif(nif, userRequest);
-                yield userResponse.map(usr -> new ResponseEntity<>(usr, HttpStatus.OK))
-                        .orElseGet(() -> new ResponseEntity<>(new UserResponse(), HttpStatus.NOT_FOUND));
-            }
-            default -> new ResponseEntity<>(new UserResponse(), HttpStatus.NOT_ACCEPTABLE);
-        };
+        if (Utils.isInvalidAcceptHeader(request)) {
+            return ResponseHandler.notAcceptableResponse();
+        }
+        Optional<UserResponse> userResponse = userService.updateUserNif(nif, userRequest);
+        return userResponse.map(usr -> ResponseHandler.buildResponse("User updated successfully", HttpStatus.OK, usr))
+                .orElseGet(() -> ResponseHandler.buildResponse("User not found: " + nif, HttpStatus.OK,
+                        new UserResponse(nif, "", "")));
     }
 
 
     @Override
-    public ResponseEntity<UserResponse> deleteUserByNif(@PathVariable("nif") String nif) {
-        String acceptStr = request.getHeader(Constants.ACCEPT_STRING);
-        return switch (acceptStr) {
-            case MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_XML_VALUE -> {
-                if (userService.deleteUserByNif(nif)) {
-                    yield new ResponseEntity<>(new UserResponse(), HttpStatus.NO_CONTENT);
-                }
-                yield new ResponseEntity<>(new UserResponse(), HttpStatus.BAD_REQUEST);
-            }
-            default -> new ResponseEntity<>(new UserResponse(), HttpStatus.NOT_ACCEPTABLE);
-        };
+    public ResponseEntity<Object> deleteUserByNif(@PathVariable("nif") String nif) {
+        if (Utils.isInvalidAcceptHeader(request)) {
+            return ResponseHandler.notAcceptableResponse();
+        }
+        if (userService.deleteUserByNif(nif)) {
+            return ResponseHandler.buildResponse("User deleted sucesfully", HttpStatus.NO_CONTENT,
+                    new UserResponse("", "", ""));
+        }
+        return ResponseHandler.buildResponse("User not found: " + nif, HttpStatus.BAD_REQUEST, nif);
     }
 }
